+++ src/layers/StaticImageLayer.js
@@ -2,6 +2,7 @@ import { CompositeLayer, COORDINATE_SYSTEM } from '@deck.gl/core';
 import XRLayer from './XRLayer';
 import { padTileWithZeros } from '../loaders/utils';
 import { NO_WEBGL2 } from '../constants';
+import { to32BitFloat } from './utils';
 
 const defaultProps = {
   pickable: true,
@@ -39,14 +40,9 @@ function scaleBounds({ width, height, translate, scale }) {
  * buffer, but without digging deeper into the WebGL it is a reasonable fix.
  */
 function padEven(data, width, height) {
-  const rgbMulitplier = NO_WEBGL2 ? 3 : 1;
   const targetWidth = (width * height) % 2 === 0 ? width : width + 1;
   const padded = data.map(d =>
-    padTileWithZeros(
-      { data: d, width: width * rgbMulitplier, height },
-      targetWidth * rgbMulitplier,
-      height
-    )
+    padTileWithZeros({ data: d, width, height }, targetWidth, height)
   );
   return { data: padded, width: targetWidth, height };
 }
@@ -70,7 +66,9 @@ export default class StaticImageLayer extends CompositeLayer {
   initializeState() {
     const { loader, z, loaderSelection } = this.props;
     loader.getRaster({ z, loaderSelection }).then(({ data, width, height }) => {
-      this.setState(padEven(data, width, height));
+      this.setState(
+        padEven(NO_WEBGL2 ? to32BitFloat(data) : data, width, height)
+      );
     });
   }
 
@@ -86,7 +84,9 @@ export default class StaticImageLayer extends CompositeLayer {
       loader
         .getRaster({ z, loaderSelection })
         .then(({ data, width, height }) => {
-          this.setState(padEven(data, width, height));
+          this.setState(
+            padEven(NO_WEBGL2 ? to32BitFloat(data) : data, width, height)
+          );
         });
     }
   }