+++ src/loaders/OMETiffLoader.js
@@ -1,11 +1,6 @@
 import OMEXML from './omeXML';
-import {
-  isInTileBounds,
-  byteSwapInplace,
-  to32BitFloat,
-  padTileWithZeros
-} from './utils';
-import { DTYPE_VALUES, NO_WEBGL2 } from '../constants';
+import { isInTileBounds, byteSwapInplace, padTileWithZeros } from './utils';
+import { DTYPE_VALUES } from '../constants';
 import { range } from '../layers/utils';
 
 const DTYPE_LOOKUP = {
@@ -179,12 +174,9 @@ export default class OMETiffLoader {
 
     const tiles = await Promise.all(tileRequests);
     return {
-      data: NO_WEBGL2 ? to32BitFloat(tiles) : tiles,
+      data: tiles,
       width: tileSize,
-      height: tileSize,
-      byteLength: NO_WEBGL2
-        ? 3 * tileSize ** 2
-        : tiles[0].BYTES_PER_ELEMENT * tileSize ** 2
+      height: tileSize
     };
   }
 
@@ -234,7 +226,7 @@ export default class OMETiffLoader {
     const width = image.getWidth();
     const height = image.getHeight();
     return {
-      data: NO_WEBGL2 ? to32BitFloat(rasters) : rasters,
+      data: rasters,
       width,
       height
     };