+++ src/loaders/utils.js
@@ -119,16 +119,3 @@ export async function getChannelStats({ loader, loaderSelection }) {
   });
   return channelStats;
 }
-
-export function to32BitFloat(data) {
-  const data32bit = data.map(arr => {
-    const { length } = arr;
-    const arr32Bits = new Float32Array(length * 3).fill(0);
-    // eslint-disable-next-line no-plusplus
-    for (let i = 0; i < length; i++) {
-      arr32Bits[i * 3] = arr[i];
-    }
-    return arr32Bits;
-  });
-  return data32bit;
-}